Remove dependence on pyspark for non-spark notebooks
